# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMC Documentation",
        "description": "Api para criaçao e gerenciamento de prontuarios, juntamente com analise de batimentos cardiacos",
        "contact": {
            "url": "https://github.com/gmatheus66"
        },
        "license": {
            "name": "License: MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://apisimc.herokuapp.com/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/doctor/detail": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "description": "Returns a detail doctor",
                "operationId": "App\\Http\\Controllers\\DoctorController::detail_auth_user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/doctor/check": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "description": "Returns a check Doctor",
                "operationId": "App\\Http\\Controllers\\DoctorController::check_user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/patient/detail": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "description": "Returns a detail Patient",
                "operationId": "App\\Http\\Controllers\\PatientController::detail_auth_user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/patient/check": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "description": "Returns a check Patient",
                "operationId": "App\\Http\\Controllers\\PatientController::check_user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/relative/register": {
            "post": {
                "tags": [
                    "Relative"
                ],
                "description": "Relative Register",
                "operationId": "App\\Http\\Controllers\\RelativeController::register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true
                    },
                    {
                        "name": "birthday",
                        "in": "query",
                        "description": "Birthday",
                        "required": true
                    },
                    {
                        "name": "sex",
                        "in": "query",
                        "description": "Sex",
                        "required": true
                    },
                    {
                        "name": "telephone",
                        "in": "query",
                        "description": "Telephone",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true
                    },
                    {
                        "name": "occupation",
                        "in": "query",
                        "description": "Occupation",
                        "required": true
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address",
                        "required": true
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City",
                        "required": true
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Country",
                        "required": true
                    },
                    {
                        "name": "state_province",
                        "in": "query",
                        "description": "State Province",
                        "required": true
                    },
                    {
                        "name": "zip",
                        "in": "query",
                        "description": "Zip",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/relative/logout": {
            "post": {
                "tags": [
                    "Relative"
                ],
                "description": "Login Relative",
                "operationId": "App\\Http\\Controllers\\RelativeController::logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/relative/login": {
            "post": {
                "tags": [
                    "Relative"
                ],
                "description": "Login Relative",
                "operationId": "App\\Http\\Controllers\\RelativeController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/relative/detail": {
            "get": {
                "tags": [
                    "Relative"
                ],
                "description": "Returns a detail Relative",
                "operationId": "App\\Http\\Controllers\\RelativeController::detail_auth_user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/relative/check": {
            "get": {
                "tags": [
                    "Relative"
                ],
                "description": "Returns a check Relative",
                "operationId": "App\\Http\\Controllers\\RelativeController::check_user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/relative/patient": {
            "get": {
                "tags": [
                    "Relative"
                ],
                "description": "Returns patients linked to family members",
                "operationId": "App\\Http\\Controllers\\RelativePatientController::get_my_patient_relatives",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/register/patient": {
            "post": {
                "tags": [
                    "Relative"
                ],
                "description": "relative Register",
                "operationId": "App\\Http\\Controllers\\RelativePatientController::register",
                "parameters": [
                    {
                        "name": "relative_id",
                        "in": "query",
                        "description": "Relative ID",
                        "required": true
                    },
                    {
                        "name": "patient_id",
                        "in": "query",
                        "description": "Patient ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}